+++ app/src/main/java/com/pos/leaders/leaderspossystem/DataBaseAdapter/ValueOfPointDB.java
@@ -8,7 +8,7 @@ import android.database.sqlite.SQLiteDatabase;
 import android.util.Log;
 
 import com.pos.leaders.leaderspossystem.DbHelper;
-import com.pos.leaders.leaderspossystem.Tools.ValueOfPoint;
+import com.pos.leaders.leaderspossystem.Models.ValueOfPoint;
 
 /**
  * Created by Win8.1 on 7/26/2017.
@@ -20,9 +20,9 @@ public class ValueOfPointDB {
     protected static final String Value_COLUMN_Id = "id";
     protected static final String Value_COLUMN = "value";
     protected static final String CreateDate_Value_COLUMN_Point = "createDate";
-    public static final String DATABASE_CREATE= "CREATE TABLE value_ofPoint ( `id` INTEGER ,"+" `value` INTEGER ,`createDate` TEXT DEFAULT current_timestamp )";
+    public static final String DATABASE_CREATE = "CREATE TABLE value_ofPoint ( `id` INTEGER ," + " `value` INTEGER ,`createDate` TEXT DEFAULT current_timestamp )";
     private DbHelper dbHelper;
-Context context;
+    Context context;
     private SQLiteDatabase db;
 
     public ValueOfPointDB(Context context) {
@@ -43,9 +43,11 @@ Context context;
     public SQLiteDatabase getDatabaseInstance() {
         return db;
     }
+
     public int insertEntry(int  id, int value,String create_date){
         ContentValues val = new ContentValues();
         //Assign values for each row.
+
         val.put(Value_COLUMN_Id,id);
         val.put(Value_COLUMN,value);
 
@@ -61,19 +63,20 @@ Context context;
             return 0;
         }
     }
+
     public ValueOfPoint getValue(){
         ValueOfPoint valueOfPoint=null;
         Cursor cursor = db.rawQuery("SELECT  * FROM " + ValueOfPoint_TABLE_NAME, null);
-        if (cursor.getCount() < 1) // UserName Not Exist
+        if (cursor.getCount() < 1) // Entry Not Exist
         {
             cursor.close();
             return valueOfPoint;
         }        cursor.moveToLast();
 
 
-        valueOfPoint =new ValueOfPoint(Integer.parseInt(cursor.getString(cursor.getColumnIndex(Value_COLUMN_Id))),Double.parseDouble(
+        valueOfPoint =new ValueOfPoint(Long.parseLong(cursor.getString(cursor.getColumnIndex(Value_COLUMN_Id))),Double.parseDouble(
                 cursor.getString(cursor.getColumnIndex(Value_COLUMN))),
                 cursor.getString(cursor.getColumnIndex(CreateDate_Value_COLUMN_Point)));
-  return valueOfPoint;
+        return valueOfPoint;
     }
 }
